refactor: miscellaneous client updates
