+++ client/package.json
@@ -41,7 +41,7 @@
 		"eslint-plugin-vue": "^9.20.1",
 		"happy-dom": "^13.1.4",
 		"jest-serializer-vue-tjw": "^3.20.0",
-		"pdap-design-system": "^3.1.0-beta.14",
+		"pdap-design-system": "^3.1.0-beta.17",
 		"postcss": "^8.4.32",
 		"tailwindcss": "^3.3.6",
 		"unplugin-vue-router": "^0.10.2",
