+++ client/src/pages/index.vue
@@ -1,3 +1,9 @@
+<route>
+	{
+		alias: '/search',
+	}
+</route>
+
 <template>
 	<main class="p-6 py-10 @container w-full overflow-scroll">
 		<h1>We help people use police data for public oversight</h1>
@@ -10,9 +16,3 @@
 <script setup>
 import SearchForm from '@/components/SearchForm.vue';
 </script>
-
-<route>
-	{
-		alias: '/search',
-	}
-</route>
