+++ client/src/pages/sign-up/success.vue
