+++ client/src/util/constants.js
@@ -1,135 +1,97 @@
 export const ALL_LOCATION_TYPES = ['locality', 'county', 'state', 'federal'];
 
-export const DATA_SOURCE_UI_SHAPE = [
-	// {
-	// 	header: 'Agency',
-	// 	records: [
-	// 		{
-	// 			title: 'Name',
-	// 			key: 'agency_name',
-	// 			// component: 'Button',
-	// 			// classNames:
-	// 			// 	'decoration-solid decoration-[6%] text-brand-gold  underline-offset-[5%] hover:brightness-85 small p-0 font-normal text-lg',
-	// 			// attributes: { intent: 'tertiary' },
-	// 			// ['data-test']: 'agency-name-button',
-	// 		},
-	// 		{ title: 'State', key: 'state_iso' },
-	// 		{
-	// 			title: 'Source URL',
-	// 			key: 'source_url',
-	// 			component: 'a',
-	// 			attributes: { target: '_blank', rel: 'noreferrer' },
-	// 			classNames: 'pdap-button-primary',
-	// 		},
-	// 		{ title: 'County', key: 'county_name' },
-	// 		{ title: 'Municipality', key: 'municipality' },
-	// 		{ title: 'Agency Type', key: 'agency_type' },
-	// 		{ title: 'Jurisdiction Type', key: 'jurisdiction_type' },
-	// 	],
-	// },
-	// {
-	// 	header: 'Data Type',
-	// 	records: [
-	// 		{
-	// 			title: 'Record type',
-	// 			key: 'record_type',
-	// 		},
-	// 		{ title: 'Description', key: 'description' },
-	// 		{ title: 'Tags', key: 'tags' },
-	// 	],
-	// },
+export const NAV_LINKS = [
 	{
-		header: 'Access & format',
-		records: [
-			{
-				title: 'Source URL',
-				key: 'source_url',
-				component: 'a',
-				attributes: { target: '_blank', rel: 'noreferrer' },
-				classNames: 'w-full inline-block truncate-text',
-			},
-			// {
-			// 	key: 'source_url_cache',
-			// 	attributes: { intent: 'secondary' },
-			// 	component: 'PButton',
-			// 	classNames: 'flex gap-4 items-center',
-			// 	['data-test']: 'view-archives-button',
-			// 	text: 'View Archives',
-			// 	renderIf: 'last_cached',
-			// 	icon: 'fa-external-link',
-			// },
-			{
-				title: 'Last Archived',
-				key: 'last_cached',
-				isDate: true,
-			},
-			{
-				title: 'ReadMe URL',
-				key: 'readme_url',
-				component: 'a',
-				classNames: 'w-full inline-block truncate-text',
-				attributes: { target: '_blank', rel: 'noreferrer' },
-			},
-			{ title: 'Access Type', key: 'access_type' },
-			{
-				title: 'Record Formats',
-				key: 'record_format',
-				component: 'span',
-				classNames:
-					'mt-1 py-[.125rem] px-3 rounded-full bg-slate-200 dark:bg-slate-600 w-fit small',
-			},
-			{ title: 'Detail Level', key: 'detail_level' },
-			{ title: 'Size', key: 'size' },
-			{ title: 'Access Notes', key: 'access_notes' },
-			// records_not_online to be hidden
-			// { title: "Records Not Online", key: "records_not_online" },
-		],
+		path: '/',
+		text: 'Search',
+		method: 'route',
 	},
 	{
-		header: 'Provenance',
-		records: [
-			{ title: 'Agency Supplied', key: 'agency_supplied' },
-			{ title: 'Supplying Entity', key: 'supplying_entity' },
-			{ title: 'Agency Originated', key: 'agency_originated' },
-			{ title: 'Originating Entity', key: 'originating_entity' },
-		],
+		href: 'https://pdap.io/data',
+		text: 'Data',
+		method: 'href',
 	},
 	{
-		header: 'Coverage & retention',
-		records: [
-			{
-				title: 'Coverage Start Date',
-				key: 'coverage_start',
-				isDate: true,
-			},
-			{ title: 'Coverage End Date', key: 'coverage_end', isDate: true },
-			{
-				title: 'Source Last Updated',
-				key: 'source_last_updated',
-				isDate: true,
-			},
-			{ title: 'Update Frequency', key: 'update_frequency' },
-			{ title: 'Update Method', key: 'update_method' },
-			{ title: 'Retention Schedule', key: 'retention_schedule' },
-			{
-				title: 'Number of Records Available',
-				key: 'number_of_records_available',
-			},
-		],
+		href: 'https://pdap.io/community',
+		text: 'Community',
+		method: 'href',
 	},
 	{
-		header: 'Data Source Meta',
-		records: [
-			{
-				title: 'Scraper URL',
-				key: 'scraper_url',
-				component: 'a',
-				attributes: { target: '_blank', rel: 'noreferrer' },
-				classNames: 'w-full inline-block truncate-text',
-			},
-			{ title: 'Created', key: 'data_source_created', isDate: true },
-			{ title: 'Agency ID', key: 'agency_id' },
-			{ title: 'Data Source ID', key: 'data_source_id' },
-		],
+		href: 'https://pdap.io/about',
+		text: 'About',
+		method: 'href',
+	},
+	{
+		href: 'https://pdap.io/donate',
+		text: 'Donate',
+		method: 'href',
+	},
+	{
+		href: 'https://docs.pdap.io/',
+		text: 'Docs',
+		method: 'href',
 	},
 ];
+
+export const STATES_TO_ABBREVIATIONS = new Map([
+	['Alabama', 'AL'],
+	['Alaska', 'AK'],
+	['American Samoa', 'AS'],
+	['Arizona', 'AZ'],
+	['Arkansas', 'AR'],
+	['Armed Forces Americas', 'AA'],
+	['Armed Forces Europe', 'AE'],
+	['Armed Forces Pacific', 'AP'],
+	['California', 'CA'],
+	['Colorado', 'CO'],
+	['Connecticut', 'CT'],
+	['Delaware', 'DE'],
+	['District Of Columbia', 'DC'],
+	['Florida', 'FL'],
+	['Georgia', 'GA'],
+	['Guam', 'GU'],
+	['Hawaii', 'HI'],
+	['Idaho', 'ID'],
+	['Illinois', 'IL'],
+	['Indiana', 'IN'],
+	['Iowa', 'IA'],
+	['Kansas', 'KS'],
+	['Kentucky', 'KY'],
+	['Louisiana', 'LA'],
+	['Maine', 'ME'],
+	['Marshall Islands', 'MH'],
+	['Maryland', 'MD'],
+	['Massachusetts', 'MA'],
+	['Michigan', 'MI'],
+	['Minnesota', 'MN'],
+	['Mississippi', 'MS'],
+	['Missouri', 'MO'],
+	['Montana', 'MT'],
+	['Nebraska', 'NE'],
+	['Nevada', 'NV'],
+	['New Hampshire', 'NH'],
+	['New Jersey', 'NJ'],
+	['New Mexico', 'NM'],
+	['New York', 'NY'],
+	['North Carolina', 'NC'],
+	['North Dakota', 'ND'],
+	['Northern Mariana Islands', 'NP'],
+	['Ohio', 'OH'],
+	['Oklahoma', 'OK'],
+	['Oregon', 'OR'],
+	['Pennsylvania', 'PA'],
+	['Puerto Rico', 'PR'],
+	['Rhode Island', 'RI'],
+	['South Carolina', 'SC'],
+	['South Dakota', 'SD'],
+	['Tennessee', 'TN'],
+	['Texas', 'TX'],
+	['US Virgin Islands', 'VI'],
+	['Utah', 'UT'],
+	['Vermont', 'VT'],
+	['Virginia', 'VA'],
+	['Washington', 'WA'],
+	['West Virginia', 'WV'],
+	['Wisconsin', 'WI'],
+	['Wyoming', 'WY'],
+]);
