--- client/src/util/links.js
@@ -1,32 +1,0 @@
-export const links = [
-	{
-		path: '/',
-		text: 'Search',
-		method: 'route',
-	},
-	{
-		href: 'https://pdap.io/data',
-		text: 'Data',
-		method: 'href',
-	},
-	{
-		href: 'https://pdap.io/community',
-		text: 'Community',
-		method: 'href',
-	},
-	{
-		href: 'https://pdap.io/about',
-		text: 'About',
-		method: 'href',
-	},
-	{
-		href: 'https://pdap.io/donate',
-		text: 'Donate',
-		method: 'href',
-	},
-	{
-		href: 'https://docs.pdap.io/',
-		text: 'Docs',
-		method: 'href',
-	},
-];
