--- client/src/util/statesToAbbreviations.js
@@ -1,62 +1,0 @@
-export default new Map([
-	['Alabama', 'AL'],
-	['Alaska', 'AK'],
-	['American Samoa', 'AS'],
-	['Arizona', 'AZ'],
-	['Arkansas', 'AR'],
-	['Armed Forces Americas', 'AA'],
-	['Armed Forces Europe', 'AE'],
-	['Armed Forces Pacific', 'AP'],
-	['California', 'CA'],
-	['Colorado', 'CO'],
-	['Connecticut', 'CT'],
-	['Delaware', 'DE'],
-	['District Of Columbia', 'DC'],
-	['Florida', 'FL'],
-	['Georgia', 'GA'],
-	['Guam', 'GU'],
-	['Hawaii', 'HI'],
-	['Idaho', 'ID'],
-	['Illinois', 'IL'],
-	['Indiana', 'IN'],
-	['Iowa', 'IA'],
-	['Kansas', 'KS'],
-	['Kentucky', 'KY'],
-	['Louisiana', 'LA'],
-	['Maine', 'ME'],
-	['Marshall Islands', 'MH'],
-	['Maryland', 'MD'],
-	['Massachusetts', 'MA'],
-	['Michigan', 'MI'],
-	['Minnesota', 'MN'],
-	['Mississippi', 'MS'],
-	['Missouri', 'MO'],
-	['Montana', 'MT'],
-	['Nebraska', 'NE'],
-	['Nevada', 'NV'],
-	['New Hampshire', 'NH'],
-	['New Jersey', 'NJ'],
-	['New Mexico', 'NM'],
-	['New York', 'NY'],
-	['North Carolina', 'NC'],
-	['North Dakota', 'ND'],
-	['Northern Mariana Islands', 'NP'],
-	['Ohio', 'OH'],
-	['Oklahoma', 'OK'],
-	['Oregon', 'OR'],
-	['Pennsylvania', 'PA'],
-	['Puerto Rico', 'PR'],
-	['Rhode Island', 'RI'],
-	['South Carolina', 'SC'],
-	['South Dakota', 'SD'],
-	['Tennessee', 'TN'],
-	['Texas', 'TX'],
-	['US Virgin Islands', 'VI'],
-	['Utah', 'UT'],
-	['Vermont', 'VT'],
-	['Virginia', 'VA'],
-	['Washington', 'WA'],
-	['West Virginia', 'WV'],
-	['Wisconsin', 'WI'],
-	['Wyoming', 'WY'],
-]);
